🛠️ chore(ErrorCode): 도메인 관련성을 고려해 별도 파일로 분리
